fix tests and clean up classes
