+++ src/main/java/uk/gov/ons/census/fwmt/csvservice/implementation/ce/CECanonicalBuilder.java
@@ -52,5 +52,4 @@ public final class CECanonicalBuilder {
 
     return createJobRequest;
   }
-
 }
