+++ src/main/java/uk/gov/ons/census/fwmt/csvservice/implementation/ce/CEConverterService.java
@@ -2,7 +2,6 @@ package uk.gov.ons.census.fwmt.csvservice.implementation.ce;
 
 import com.opencsv.bean.CsvToBean;
 import com.opencsv.bean.CsvToBeanBuilder;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.Resource;
