+++ src/main/java/uk/gov/ons/census/fwmt/csvservice/message/GatewayActionProducer.java
@@ -2,7 +2,6 @@ package uk.gov.ons.census.fwmt.csvservice.message;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.amqp.core.DirectExchange;
 import org.springframework.amqp.core.Message;
 import org.springframework.amqp.core.MessageProperties;
