+++ src/test/java/uk/gov/ons/fwmt/csvservice/implementation/ccs/CCSCanonicalBuilderTest.java
@@ -1,37 +1,18 @@
-package uk.gov.ons.fwmt.csvservice.canonical;
+package uk.gov.ons.fwmt.csvservice.implementation.ccs;
 
 import org.junit.Test;
 import uk.gov.ons.census.fwmt.canonical.v1.CreateFieldWorkerJobRequest;
 import uk.gov.ons.census.fwmt.csvservice.dto.CCSPropertyListing;
-import uk.gov.ons.census.fwmt.csvservice.dto.CEJobListing;
 import uk.gov.ons.fwmt.csvservice.helper.CSVRecordDTOBuilder;
-
-import java.util.UUID;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static uk.gov.ons.census.fwmt.csvservice.implementation.ccs.CCSCanonicalBuilder.createCCSJob;
-import static uk.gov.ons.census.fwmt.csvservice.implementation.ce.CECanonicalBuilder.createCEJob;
 
-public class CanonicalJobHelperTest {
-
-  @Test
-  public void createCE() {
-    // Given
-    CEJobListing ceJobListing = new CSVRecordDTOBuilder().createCECSVRecord();
-
-    // When
-    CreateFieldWorkerJobRequest createFieldWorkerJobRequest = createCEJob(ceJobListing);
-
-    // Then
-    assertEquals(UUID.fromString(ceJobListing.getCaseId()), createFieldWorkerJobRequest.getCaseId());
-    assertEquals(ceJobListing.getEstablishmentType(), createFieldWorkerJobRequest.getEstablishmentType());
-    assertEquals(ceJobListing.getLatitude(), createFieldWorkerJobRequest.getAddress().getLatitude());
-    assertEquals(ceJobListing.getCoordinatorId(), createFieldWorkerJobRequest.getCoordinatorId());
-  }
+public class CCSCanonicalBuilderTest {
 
   @Test
-  public void createCCS() {
+  public void createCCS () {
     // Given
     CCSPropertyListing ccsPropertyListing = new CSVRecordDTOBuilder().createCCSCSVRecord();
