+++ src/test/java/uk/gov/ons/fwmt/csvservice/implementation/ce/CECanonicalBuilderTest.java
@@ -1,22 +1,20 @@
-package uk.gov.ons.fwmt.csvservice.canonical;
+package uk.gov.ons.fwmt.csvservice.implementation.ce;
 
 import org.junit.Test;
 import uk.gov.ons.census.fwmt.canonical.v1.CreateFieldWorkerJobRequest;
-import uk.gov.ons.census.fwmt.csvservice.dto.CCSPropertyListing;
 import uk.gov.ons.census.fwmt.csvservice.dto.CEJobListing;
 import uk.gov.ons.fwmt.csvservice.helper.CSVRecordDTOBuilder;
 
 import java.util.UUID;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static uk.gov.ons.census.fwmt.csvservice.implementation.ccs.CCSCanonicalBuilder.createCCSJob;
 import static uk.gov.ons.census.fwmt.csvservice.implementation.ce.CECanonicalBuilder.createCEJob;
 
-public class CanonicalJobHelperTest {
+
+public class CECanonicalBuilderTest {
 
   @Test
-  public void createCE() {
+  public void createCE () {
     // Given
     CEJobListing ceJobListing = new CSVRecordDTOBuilder().createCECSVRecord();
 
@@ -28,20 +26,5 @@ public class CanonicalJobHelperTest {
     assertEquals(ceJobListing.getEstablishmentType(), createFieldWorkerJobRequest.getEstablishmentType());
     assertEquals(ceJobListing.getLatitude(), createFieldWorkerJobRequest.getAddress().getLatitude());
     assertEquals(ceJobListing.getCoordinatorId(), createFieldWorkerJobRequest.getCoordinatorId());
-  }
-
-  @Test
-  public void createCCS() {
-    // Given
-    CCSPropertyListing ccsPropertyListing = new CSVRecordDTOBuilder().createCCSCSVRecord();
-
-    // When
-    CreateFieldWorkerJobRequest createFieldWorkerJobRequest = createCCSJob(ccsPropertyListing);
-
-    // Then
-    assertNotNull(createFieldWorkerJobRequest.getCaseId());
-    assertEquals(ccsPropertyListing.getPostCode(), createFieldWorkerJobRequest.getAddress().getPostCode());
-    assertEquals(ccsPropertyListing.getLatitude(), createFieldWorkerJobRequest.getAddress().getLatitude());
-    assertEquals(ccsPropertyListing.getCoordinatorId(), createFieldWorkerJobRequest.getCoordinatorId());
   }
 }
