+++ src/test/java/uk/gov/ons/fwmt/csvservice/implementation/ce/CEConverterServiceImplTest.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.fwmt.csvservice.service;
+package uk.gov.ons.fwmt.csvservice.implementation.ce;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -11,8 +11,8 @@ import org.springframework.core.io.Resource;
 import org.springframework.core.io.WritableResource;
 import org.springframework.test.util.ReflectionTestUtils;
 import uk.gov.ons.census.fwmt.common.error.GatewayException;
-import uk.gov.ons.census.fwmt.csvservice.implementation.ce.CCSConverterServiceImpl;
-import uk.gov.ons.census.fwmt.csvservice.implementation.ce.GatewayActionAdapter;
+import uk.gov.ons.census.fwmt.csvservice.adapter.GatewayActionAdapter;
+import uk.gov.ons.census.fwmt.csvservice.implementation.ce.CEConverterService;
 import uk.gov.ons.census.fwmt.events.component.GatewayEventManager;
 
 import java.io.OutputStream;
@@ -20,13 +20,13 @@ import java.nio.file.Path;
 
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
-import static uk.gov.ons.census.fwmt.csvservice.config.GatewayEventsConfig.CSV_CCS_REQUEST_EXTRACTED;
+import static uk.gov.ons.census.fwmt.csvservice.implementation.ce.CEGatewayEventsConfig.CSV_CE_REQUEST_EXTRACTED;
 
 @RunWith(MockitoJUnitRunner.class)
-public class CCSConverterServiceImplTest {
+public class CEConverterServiceImplTest {
 
   @InjectMocks
-  private CCSConverterService ccsConverterService;
+  private CEConverterService ceConverterService;
 
   @Mock
   private GatewayActionAdapter gatewayActionAdapter;
@@ -41,21 +41,21 @@ public class CCSConverterServiceImplTest {
   private WritableResource writableResource;
 
   @Test
-  public void convertCCSCSVToCanonicalTest() throws GatewayException {
+  public void convertCECSVToCanonicalTest() throws GatewayException {
     // Given
     ClassLoader classLoader = getClass().getClassLoader();
-    String testPathString = classLoader.getResource("testCCSCSV.csv").getPath();
+    String testPathString = classLoader.getResource("testCECSV.csv").getPath();
     Path testPath = Path.of("/", testPathString);
     Resource testResource = new FileSystemResource(testPathString);
 
-    ReflectionTestUtils.setField(ccsConverterService, "csvGCPFile", testResource);
-    ReflectionTestUtils.setField(ccsConverterService, "csvPath", testPath);
-    ReflectionTestUtils.setField(ccsConverterService, "processedPath", testPath);
+    ReflectionTestUtils.setField(ceConverterService, "csvGCPFile", testResource);
+    ReflectionTestUtils.setField(ceConverterService, "csvPath", testPath);
+    ReflectionTestUtils.setField(ceConverterService, "processedPath", testPath);
 
     // When
-    ccsConverterService.convertCSVToCanonical();
+    ceConverterService.convertCSVToCanonical();
 
     // Then
-    Mockito.verify(gatewayEventManager).triggerEvent(anyString(), eq(CSV_CCS_REQUEST_EXTRACTED));
+    Mockito.verify(gatewayEventManager).triggerEvent(anyString(), eq(CSV_CE_REQUEST_EXTRACTED));
   }
 }
